// frontend/src/api/model/pageInfoMovie.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Movie Review System API
 * Backend API for Movie Review Website
 * OpenAPI spec version: v1.0
 */
import type { Movie } from "./movie";

export interface PageInfoMovie {
  total?: number;
  list?: Movie[];
  pageNum?: number;
  pageSize?: number;
  size?: number;
  startRow?: number;
  endRow?: number;
  pages?: number;
  prePage?: number;
  nextPage?: number;
  isFirstPage?: boolean;
  isLastPage?: boolean;
  hasPreviousPage?: boolean;
  hasNextPage?: boolean;
  navigatePages?: number;
  navigatepageNums?: number[];
  navigateFirstPage?: number;
  navigateLastPage?: number;
}